% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bootstrap.R
\name{bootstrap_model}
\alias{bootstrap_model}
\title{Fitting the HDMVMediation model}
\usage{
bootstrap_model(
  mediators,
  confounders,
  trt,
  outcomes,
  quiet_msglasso = TRUE,
  lam1.v = seq(0.001, 0.05, length = 20),
  lamG.v = seq(0.001, 0.05, length = 20),
  alpha = 0.05,
  nB = 5000,
  msg_folds = 5,
  seed = 823543,
  outcome_grps = FALSE,
  NumOutGrps = NULL,
  OutGrpStarts = NULL,
  OutGrpEnds = NULL
)
}
\arguments{
\item{mediators}{An \eqn{n\times p} numeric matrix of candidate mediators.}

\item{confounders}{An \eqn{n\times\ell} numeric matrix of confounders.}

\item{trt}{An \eqn{n\times 1} numeric matrix or vector of treatment/exposure status.}

\item{outcomes}{An \eqn{n\times q} numeric matrix of outcomes.}

\item{quiet_msglasso}{A boolean indicator determining whether or not the CV progress of the MSGLasso estimation should be suppressed; defaults to `TRUE`.}

\item{lam1.v}{A numeric vector of component-wise regularization parameters as in the MSGLasso of Li et al. (2015). By default, this is taken to be `seq(from = 1e-03, to = 0.05, length = 20)`.}

\item{lamG.v}{A numeric vector of group-wise regularization parameters as in the MSGLasso of Li et al. (2015). By default, this is taken to be `seq(from = 1e-03, to = 0.05, length = 20)`.}

\item{alpha}{A numeric between 0 and 1 denoting the confidence level for inference; defaults to 0.05.}

\item{nB}{A numeric determining the number of bootstrap samples taken for inference. By default, 5000 samples will be used.}

\item{msg_folds}{A numeric determining the number of CV folds to be used in tuning the MSGLasso regularization parameters. By default, this is taken to be 5.}

\item{seed}{An integer seed for reproducible bootstrap inference.}

\item{outcome_grps}{A boolean indicator for whether or not there are any groups on the outcome variables. By default, there are no outcome groups so this argument is `FALSE`.}

\item{NumOutGrps}{An integer denoting the number of outcome groups; default is `NULL` in keeping with the default `outcome_grps = FALSE`}

\item{OutGrpStarts}{A vector of starting coordinates for the outcome groups. Equivalent to `R.Starts` from the `MSGLasso` package.}

\item{OutGrpEnds}{A vector of ending coordinates for the outcome groups. Equivalent to `R.Ends` from the `MSGLasso` package.}
}
\value{
A \eqn{(p*q+2*q)\times 8} matrix `mod_boot_summ` with columns:
* `OrigEst`: the original PIDE/TIDE/DE estimate;
* `Mean_boot`: the mean of the bootstrap PIDE/TIDE/DE estimates;
* `boot_SE`: the standard deviation of the bootstrap PIDE/TIDE/DE estimates;
* `boot_pval`: the unadjusted p-value for the bootstrap PIDE/TIDE/DE estimates;
* `bca_lowerCL`: the lower limit of the \eqn{100(1-\texttt{alpha})\%} BCa CI;
* `bca_upperCL`: the upper limit of the \eqn{100(1-\texttt{alpha})\%} BCa CI;
* `per_lowerCL`: the lower limit of the \eqn{100(1-\texttt{alpha})\%} percentile CI;
* `per_upperCL`: the upper limit of the \eqn{100(1-\texttt{alpha})\%} percentile CI.
}
\description{
`bootstrap_model` is the main function for fitting and performing inference for the high-dimensional multivariate outcomes mediation model of Sun, Xiao, and Wu (2025).
}
\examples{
## Load toy data
data(hdmvmed_test_data)

## Assign mediator/confounder/trt/outcome matrices
mediators <- hdmvmed_test_data[,7:106]
confounders <- hdmvmed_test_data[,107:111]
trt <- hdmvmed_test_data[,1]
outcomes <- hdmvmed_test_data[,2:6]

## Fit the model
model_fit <- bootstrap_model(mediators = mediators, confounders = confounders,
                             trt = trt, outcomes = outcomes, nB = 10)
model_fit

}
\references{
{
Li, Y., Nan, B., and Zhu, J. (2015). Multivariate Sparse Group Lasso for the 
Multivariate Multiple Linear Regression with an Arbitrary Group Structure. 
\emph{Biometrics}, *71*, 354-363.
}

{
Sun, E., Xiao, J., and Wu, T. T. (2025). Causal Mediation Analysis for Multiple 
Outcomes and High-dimensional Mediators: Identification, Inference, and Application.
\emph{Biometrics}. _Under Review._
}
}
