% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdmvm_table.R
\name{hdmvm_table}
\alias{hdmvm_table}
\title{Summary table for HDMVMediation models}
\usage{
hdmvm_table(mod_boot_summ, p, outcomes)
}
\arguments{
\item{mod_boot_summ}{A numeric matrix; the output from `bootstrap_model()`}

\item{p}{An integer representing the number of candidate mediators}

\item{outcomes}{A character vector with the names of each outcome}
}
\value{
An html widget of class `datatables`.
}
\description{
`hdmvm_table()` converts the matrix output of the main `bootstrap_model()` function into a searchable/filterable html table using the `DT` package.
}
\examples{
## Load toy data
data(hdmvmed_test_data)

## Assign mediator/confounder/trt/outcome matrices
mediators <- hdmvmed_test_data[,7:106]
confounders <- hdmvmed_test_data[,107:111]
trt <- hdmvmed_test_data[,1]
outcomes <- hdmvmed_test_data[,2:6]

## Fit the model
model_fit <- bootstrap_model(mediators = mediators, confounders = confounders,
                             trt = trt, outcomes = outcomes, nB = 10)

## Generate summary table
hdmvm_table(model_fit, p = ncol(mediators), outcomes = colnames(outcomes))

}
