\name{hdmvmed_test_data}
\alias{hdmvmed_test_data}
\docType{data}
\title{
Test Data
}
\description{
Contains toy dataset of 250 observations with: one binary treatment, A; five outcomes, Y1-Y5;
100 normally distributed mediators, M1-M100; 10 confounders, U1-U10. By response, the true TIDEs
are (25, 30, -35, 30, 25); the true DEs are (6, 9, -3, 15, 18); the true PIDEs for Y1 are
(9.19, 1.91, 4.09, 4.41, 5.39, 0, ..., 0); the true PIDEs for Y2 are (0.58, 5.48, 0.31, 11.25, 12.38, 0, ..., 0); 
the true PIDEs for Y3 are (-7.07, -5.10, -17.98, -5.44, 0.59, 0, ..., 0); the true PIDEs for Y4 
are (6.06, 1.35, 5.88, 1.65, 15.05, 0, ..., 0); and the true PIDEs for Y5 
are (13.44, 1.13, 5.14, 5.56, -0.27, 0, ..., 0).
}
\usage{data(hdmvmed_test_data)}
\keyword{datasets}
